refactor: transfer model classes into core module
