+++ app/src/main/java/com/example/tsundokun/data/repository/CategoryRepository.kt
@@ -1,6 +1,6 @@
 package com.example.tsundokun.data.repository
 
-import com.example.tsundokun.domain.models.Category
+import com.example.tsundokun.core.model.Category
 import kotlinx.coroutines.flow.Flow
 
 interface CategoryRepository {
