+++ app/src/main/java/com/example/tsundokun/ui/confirm/ConfirmScreen.kt
@@ -39,7 +39,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.tsundokun.R.string
-import com.example.tsundokun.domain.models.Category
+import com.example.tsundokun.core.model.Category
 import com.example.tsundokun.ui.confirm.ConfirmViewModel.ConfirmUiState
 import com.example.tsundokun.ui.confirm.component.AllInputFields
 import com.example.tsundokun.ui.confirm.component.data.ConfirmScreenNavArgs
