+++ app/src/main/java/com/example/tsundokun/ui/confirm/component/AllInputFields.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.example.tsundokun.R.string
-import com.example.tsundokun.domain.models.Category
+import com.example.tsundokun.core.model.Category
 import com.example.tsundokun.ui.confirm.component.jsoup.FetchHtml
 import com.example.tsundokun.ui.confirm.component.jsoup.GetTitle
