+++ core/model/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    alias(libs.plugins.android.library)
+    alias(libs.plugins.kotlin.android)
+}
+
+android {
+    namespace = "com.example.tsundokun.core.model"
+    compileSdk = 33
+
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
+}
