+++ core/model/src/main/java/com/example/tsundokun/core/model/Category.kt
@@ -1,4 +1,4 @@
-package com.example.tsundokun.domain.models
+package com.example.tsundokun.core.model
 
 data class Category(
     val id: String = "",
